src/test/java/com/leetcode/algorithm/easy/isomorphicstrings/SolutionTest.java
package com.leetcode.algorithm.easy.isomorphicstrings;

import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.assertEquals;

public class SolutionTest {
  @Test
  public void testCase1() {
    Solution solution = new Solution();
    assertEquals(true, solution.isIsomorphic("egg", "add"));
  }

  @Test
  public void testCase2() {
    Solution solution = new Solution();
    assertEquals(false, solution.isIsomorphic("foo", "bar"));
  }

  @Test
  public void testCase3() {
    Solution solution = new Solution();
    assertEquals(true, solution.isIsomorphic("paper", "title"));
  }
}
